Fix priority inversion when high-priority event is activated unexpectedly
If we activate a high-priority event while we are processing a big
queue of low-priority events, we would previously finish processing
all events at that priority before calling the dispatch function
again.  Now, any time an event is activated during event processing,
and that event is more high-priority than the currently running
events, we launch a rescan immediately.